更新代码编辑器,增加点赞

+++ src/main/java/com/yang/blog/util/FileUtils.java
@@ -332,8 +332,10 @@ public class FileUtils {
 
     public static void delete(String fileId) {
         com.yang.blog.entity.File file = ServiceConfig.serviceConfig.fileService.getById(fileId);
-        java.io.File uploadFile = new java.io.File(FileUtils.uploadPath(file));
-        uploadFile.delete();
-        ServiceConfig.serviceConfig.fileService.removeById(file.getId());
+        if (file != null) {
+            java.io.File uploadFile = new java.io.File(FileUtils.uploadPath(file));
+            uploadFile.delete();
+            ServiceConfig.serviceConfig.fileService.removeById(file.getId());
+        }
     }
 }
